requirements.txt
fastapi
uvicorn[standard]
pydantic
aiohttp
opencv-python-headless
langchain
langsmith
python-dotenv
requests
aiofiles
